src/requirements.txt
Django==3.1
psycopg2-binary>=2.8
drf-yasg==1.20.0
django-crum==0.7.9
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
django-admin-interface==0.14.2